requirements.txt
# Core dependencies
flask>=2.0.1
flask-cors>=3.0.10
tensorflow>=2.6.0
numpy>=1.19.5
pandas>=1.3.0
librosa>=0.8.1
scikit-learn>=0.24.2
pydub>=0.25.1
matplotlib>=3.4.2
pyaudio>=0.2.11
pydot>=1.4.2
graphviz>=0.16
soundfile>=0.10.3

# Additional dependencies
scipy>=1.7.0
seaborn>=0.11.1
tqdm>=4.61.2
pyyaml>=5.4.1
joblib>=1.0.1
werkzeug>=2.0.1
gunicorn>=20.1.0
numba>=0.54.0
audiomentations>=0.20.0